website now builds with diagrams-svg (99%)
All you have to do is rename web/gallery/KnightTour.lhs so it doesn't
try to build it (it contains an embedded image so SVG can't handle
it).  Then just do the usual incantations with Shake.  It will
auto-detect the backend to use based on whether the diagrams-cairo
package is installed.