updates to quickstart tutorial
This brings the tutorial up-to-date.  But there's a lot more we could
add to the tutorial as well.  It might be nice to add a section at the
end giving a somewhat realistic example diagram as a goal and then
working through how to construct it, explaining things and linking to
further information as they come up.